+++ lib/models/item.dart
@@ -1,0 +1,6 @@
+class Item {
+  final String name;
+  final String category;
+
+  Item({required this.name, required this.category});
+}
