+++ lib/screens/filter_screen.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class FilterScreen extends StatelessWidget {
+  const FilterScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Filter'),
+      ),
+      body: const Center(
+        child: Text('Filter Screen'),
+      ),
+    );
+  }
+}
