+++ lib/utils/constants.dart
@@ -1,0 +1,3 @@
+class Constants {
+  static const String appTitle = 'Flutter App';
+}
