+++ lib/widgets/item_card.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+import '../models/item.dart';
+
+class ItemCard extends StatelessWidget {
+  final Item item;
+
+  const ItemCard({super.key, required this.item});
+
+  @override
+  Widget build(BuildContext context) {
+    return Card(
+      child: ListTile(
+        title: Text(item.name),
+        subtitle: Text(item.category),
+      ),
+    );
+  }
+}
