+++ pubspec.yaml
@@ -33,7 +33,6 @@ dependencies:
     sdk: flutter
   file_picker: ^4.6.1
   csv: ^5.0.1
-  path_provider: ^2.0.10
 
 
   # The following adds the Cupertino Icons font to your application.
